x86 ld script: change text section start to 0x1000
It used to be 0x0, which means calling null pointer executed some random
function (usually main).

Writing to NULL pointer corrupts said function, too.

A better option would be to disallow access to 0x0 completely,
but unfortunately, it seems to be impossible with segmentation alone
(one need paging for that).

Minor drawback is that a 0x1000 bytes are wasted. 0x1000 might be
a bit overkill, though.