Add 'NO_DEPS' option for TemplateRender() method.
When given option is set in the scons environment, .deps files are not generated.
This may be usefull when it is known that only single template is used and target file is created in source tree.

template_generation.BuildAsmOffsets(): When parse intermediate asm file, take into account that for some languages (e.g. x86)
gcc prepends constant with "$". This symbol is discarded from generated header asm-compatible file.